Restructure APIs docs and apply PR review feedback
- Remove deprecated Scale plan from Billing (page, nav, card); redirect old platform path to Pro
- Regroup Billing: move the payments page out of Plans into 'Manage billing'
- Fold former API Integration content into the APIs section; drop the standalone section
- Add an 'Events & Errors' subsection (events, webhooks, response codes, error handling) via URL-invisible route groups; leave REST/GraphQL/Realtime untouched
- Move Release policy to Advanced
- Repoint redirects and internal links; update STYLE.md
- Add Sept 1, 2026 deprecation date to Dev keys note; fix malformed release-policy anchor
