+++ src/routes/docs/advanced/billing/payments/+page.markdoc
@@ -1,6 +1,6 @@
 ---
 layout: article
-title: Billing
+title: Manage billing
 description: Understand Appwrite's billing features, like budget caps, billing periods, taxes, and more.
 ---
