--- src/routes/docs/advanced/billing/scale/+page.markdoc
@@ -1,9 +1,0 @@
----
-layout: article
-title: Scale
-description: Appwrite's Scale plan fully supports scaling your application.
----
-
-Appwrite's Scale plan is designed for growing development teams and agencies with many organizational members and large projects.
-The plan offers unlimited seats across your organization and dedicated resources per project.
-Scale plan organizations will receive additional compliance measures, organization roles, and dedicated support.
