+++ src/routes/docs/apis/(events-and-errors)/webhooks/+page.markdoc
@@ -1474,7 +1474,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
 # Payload {% #payload %}
 
-Each event type has a specific payload format with the relevant event information. All event payloads mirror the payloads for the API payload which parallel to the [event types](/docs/advanced/api-integration/events).
+Each event type has a specific payload format with the relevant event information. All event payloads mirror the payloads for the API payload which parallel to the [event types](/docs/apis/events).
 
 # Headers {% #headers %}
 
@@ -1484,7 +1484,7 @@ HTTP requests made to your webhook's configured URL endpoint will contain severa
 |--------|-------------|
 | X-Appwrite-Webhook-Id | The ID of the Webhook who triggered the event. |
 | X-Appwrite-Webhook-Events | Names of the events that triggered this delivery. |
-| X-Appwrite-Webhook-Name | Name of the webhook as specified in your app settings and [events list](/docs/advanced/api-integration/events). |
+| X-Appwrite-Webhook-Name | Name of the webhook as specified in your app settings and [events list](/docs/apis/events). |
 | X-Appwrite-Webhook-User-Id | The user ID of the user who triggered the event. Returns an empty string if an API key triggered the event. Note that events like `account.create` or `account.sessions.create` are performed by guest users and will not return any user ID. If you still need the user ID for these events, you can find it in the event payload. |
 | X-Appwrite-Webhook-Project-Id | The ID of the project who owns the Webhook and API call. |
 | X-Appwrite-Webhook-Signature | The HMAC-SHA1 signature of the payload. This is used to verify the authenticity of the payload. |
@@ -1704,4 +1704,4 @@ You can specify one or many events to subscribe to with webhooks.
 {% /accordion_item %}
 {% /accordion %}
 
-[Learn more about events](/docs/advanced/api-integration/events)
+[Learn more about events](/docs/apis/events)
