+++ src/routes/docs/apis/realtime/+page.markdoc
@@ -116,7 +116,7 @@ To subscribe to updates from different Appwrite resources, you need to specify o
 
 If you subscribe to a channel, you will receive callbacks for a variety of events related to the channel. The events column in the callback can be used to filter and respond to specific events in a channel.
 
-[View a list of all available events](/docs/advanced/api-integration/events).
+[View a list of all available events](/docs/apis/events).
 
 {% info title="Permissions" %}
 All subscriptions are secured by the [permissions system](/docs/advanced/security/permissions) offered by Appwrite, meaning a user will only receive updates to resources they have permission to access.
