+++ src/routes/docs/references/+page.markdoc
@@ -70,4 +70,4 @@ Track which signed-in users are active right now and broadcast their status in r
 
 ## Error handling {% #error-handling %}
 
-When building with Appwrite, implement proper error handling to provide user-friendly messages instead of exposing raw error responses. For implementation details and best practices, refer to our [Error handling guide](/docs/advanced/api-integration/error-handling) and [Response codes](/docs/advanced/api-integration/response-codes#error-types) documentation.
+When building with Appwrite, implement proper error handling to provide user-friendly messages instead of exposing raw error responses. For implementation details and best practices, refer to our [Error handling guide](/docs/apis/error-handling) and [Response codes](/docs/apis/response-codes#error-types) documentation.
